30/06/2024

+++ back-end/database/migrations/2024_01_13_030718_create_orders_table.php
@@ -16,7 +16,7 @@ class CreateOrdersTable extends Migration
         Schema::create('orders', function (Blueprint $table) {
             $table->bigIncrements('order_id');
             $table->unsignedBigInteger('user_id');
-            $table->decimal('total_amount', 10, 2);
+            $table->unsignedBigInteger('total_amount');
             $table->enum('payment_status', ['paid', 'unpaid'])->default('unpaid');
             $table->date('order_date');
             $table->string('service');
